refactor: Remove dead code - prior sampling MCTS and MCTS payout loss
Remove two unused code paths that were superseded by better approaches:

1. Remove `loss_variant: mcts` training option
   - Was: full 5x5 payout matrix supervision from MCTS estimates
   - Now: always use sparse DQN loss (actual game outcomes at played action)
   - Simplified compute_losses() in all 3 training modules

2. Remove prior sampling MCTS (search.py)
   - Was: sample actions from NN policy priors during search
   - Now: DecoupledPUCTSearch is the only MCTS variant
   - Moved SearchResult dataclass to decoupled_puct.py
   - Simplified MCTSConfig to just DecoupledPUCTConfig

Files deleted:
- alpharat/mcts/search.py
- tests/mcts/test_search.py
- tests/mcts/test_scripted_search.py

Closes #31
