+++ alpharat/mcts/__init__.py
@@ -1,25 +1,26 @@
 """Monte Carlo Tree Search for simultaneous-move games."""
 
-from alpharat.mcts.decoupled_puct import DecoupledPUCTConfig, DecoupledPUCTSearch
+from alpharat.mcts.decoupled_puct import (
+    DecoupledPUCTConfig,
+    DecoupledPUCTSearch,
+    SearchResult,
+)
 from alpharat.mcts.nash import (
     compute_nash_equilibrium,
     compute_nash_value,
     select_action_from_strategy,
 )
 from alpharat.mcts.node import MCTSNode
-from alpharat.mcts.search import MCTSSearch, PriorSamplingConfig, SearchResult
 from alpharat.mcts.tree import MCTSTree
 
-MCTSConfig = PriorSamplingConfig | DecoupledPUCTConfig
+MCTSConfig = DecoupledPUCTConfig
 
 __all__ = [
     "DecoupledPUCTConfig",
     "DecoupledPUCTSearch",
     "MCTSConfig",
     "MCTSNode",
-    "MCTSSearch",
     "MCTSTree",
-    "PriorSamplingConfig",
     "SearchResult",
     "compute_nash_equilibrium",
     "compute_nash_value",
