+++ alpharat/nn/symmetric_training.py
@@ -3,10 +3,6 @@
 Similar to training.py but uses SymmetricMLP which has structural symmetry,
 so no player swap augmentation is needed.
 
-Loss variants:
-- "mcts": Full payout matrix supervision (MSE on all 25 entries from MCTS estimates)
-- "dqn": Sparse supervision (MSE only on played action pair vs actual outcome)
-
 Usage:
     from alpharat.nn.symmetric_training import SymmetricTrainConfig, run_symmetric_training
 
@@ -62,7 +58,6 @@ class SymmetricOptimConfig(BaseModel):
     nash_weight: float = 0.0  # Nash consistency loss weight (0 = disabled)
     nash_mode: Literal["target", "predicted"] = "target"
     constant_sum_weight: float = 0.0  # Constant-sum regularization weight (0 = disabled)
-    loss_variant: Literal["mcts", "dqn"] = "dqn"
     batch_size: int = 4096
 
 
@@ -135,7 +130,6 @@ def sparse_payout_loss(
 def compute_losses(
     model: SymmetricMLP,
     data: dict[str, torch.Tensor],
-    loss_variant: Literal["mcts", "dqn"],
     policy_weight: float,
     value_weight: float,
     nash_weight: float = 0.0,
@@ -153,17 +147,14 @@ def compute_losses(
     loss_p1 = F.cross_entropy(logits_p1, data["policy_p1"])
     loss_p2 = F.cross_entropy(logits_p2, data["policy_p2"])
 
-    if loss_variant == "mcts":
-        loss_value = F.mse_loss(pred_payout, data["payout_matrix"])
-    else:
-        # Sparse DQN: supervise with actual game outcomes
-        loss_value = sparse_payout_loss(
-            pred_payout,
-            data["action_p1"],
-            data["action_p2"],
-            data["p1_value"],
-            data["p2_value"],
-        )
+    # Sparse loss: supervise with actual game outcomes at played action pair
+    loss_value = sparse_payout_loss(
+        pred_payout,
+        data["action_p1"],
+        data["action_p2"],
+        data["p1_value"],
+        data["p2_value"],
+    )
 
     # Nash consistency loss (game-theoretic constraint)
     if nash_weight > 0:
@@ -416,7 +407,6 @@ def run_symmetric_training(
             out = compute_losses(
                 model,
                 batch,
-                optim_cfg.loss_variant,
                 optim_cfg.policy_weight,
                 optim_cfg.value_weight,
                 optim_cfg.nash_weight,
@@ -477,7 +467,6 @@ def run_symmetric_training(
                 vl_out = compute_losses(
                     model,
                     val_batch,
-                    optim_cfg.loss_variant,
                     optim_cfg.policy_weight,
                     optim_cfg.value_weight,
                     optim_cfg.nash_weight,
