+++ configs/tournament_test.yaml
@@ -1,8 +1,5 @@
 # Small test tournament for quick verification
 agents:
-  ps_50:
-    variant: prior_sampling
-    simulations: 50
   puct_50:
     variant: decoupled_puct
     simulations: 50
@@ -13,9 +10,10 @@ agents:
     simulations: 300
     c_puct: 1.5
 
-  ps_300:
-    variant: prior_sampling
-    simulations: 300
+  puct_1000:
+    variant: decoupled_puct
+    simulations: 1000
+    c_puct: 1.5
 
 games_per_matchup: 10
