+++ configs/train.yaml
@@ -14,7 +14,6 @@ optim:
   # - nash_mode: "target" uses MCTS policies, "predicted" uses NN's own policies
   nash_weight: 0.0
   nash_mode: target
-  loss_variant: dqn  # "mcts" (full matrix) or "dqn" (sparse)
   p_augment: 0.5      # probability of player swap augmentation per sample
 
 data:
