+++ configs/train_local_value.yaml
@@ -10,7 +10,6 @@ optim:
   policy_weight: 1.0
   value_weight: 1.0
   ownership_weight: 1.0
-  loss_variant: dqn
   p_augment: 0.5
 
 data:
