+++ tests/data/test_batch.py
@@ -16,26 +16,11 @@ from alpharat.data.batch import (
     load_batch_metadata,
     save_batch_metadata,
 )
-from alpharat.mcts import DecoupledPUCTConfig, PriorSamplingConfig
+from alpharat.mcts import DecoupledPUCTConfig
 
 
 class TestMCTSConfig:
-    """Tests for MCTS config discriminated union."""
-
-    def test_prior_sampling_from_dict(self) -> None:
-        """PriorSamplingConfig parses from dict."""
-        data = {"variant": "prior_sampling", "simulations": 800, "gamma": 0.95}
-        config = PriorSamplingConfig.model_validate(data)
-
-        assert config.variant == "prior_sampling"
-        assert config.simulations == 800
-        assert config.gamma == 0.95
-
-    def test_prior_sampling_defaults(self) -> None:
-        """PriorSamplingConfig uses default gamma."""
-        config = PriorSamplingConfig(simulations=100)
-
-        assert config.gamma == 1.0
+    """Tests for MCTS config."""
 
     def test_decoupled_puct_from_dict(self) -> None:
         """DecoupledPUCTConfig parses from dict."""
@@ -59,22 +44,8 @@ class TestMCTSConfig:
         assert config.gamma == 1.0
         assert config.c_puct == 1.5
 
-    def test_discriminated_union_prior_sampling(self) -> None:
-        """BatchMetadata parses prior_sampling via discriminator."""
-        data = {
-            "batch_id": "test-id",
-            "created_at": "2024-01-01T00:00:00Z",
-            "checkpoint_path": None,
-            "mcts_config": {"variant": "prior_sampling", "simulations": 800},
-            "game_params": {"width": 10, "height": 10, "max_turns": 200, "cheese_count": 21},
-        }
-        metadata = BatchMetadata.model_validate(data)
-
-        assert isinstance(metadata.mcts_config, PriorSamplingConfig)
-        assert metadata.mcts_config.simulations == 800
-
-    def test_discriminated_union_decoupled_puct(self) -> None:
-        """BatchMetadata parses decoupled_puct via discriminator."""
+    def test_batch_metadata_parses_config(self) -> None:
+        """BatchMetadata parses decoupled_puct config."""
         data = {
             "batch_id": "test-id",
             "created_at": "2024-01-01T00:00:00Z",
@@ -97,7 +68,7 @@ class TestBatchMetadata:
             batch_id="test",
             created_at=datetime.now(UTC),
             checkpoint_path=None,
-            mcts_config=PriorSamplingConfig(simulations=100),
+            mcts_config=DecoupledPUCTConfig(simulations=100),
             game_params=GameParams(width=10, height=10, max_turns=200, cheese_count=21),
         )
 
@@ -109,7 +80,7 @@ class TestBatchMetadata:
             batch_id="test",
             created_at=datetime.now(UTC),
             checkpoint_path="/models/checkpoint_100.pt",
-            mcts_config=PriorSamplingConfig(simulations=100),
+            mcts_config=DecoupledPUCTConfig(simulations=100),
             game_params=GameParams(width=10, height=10, max_turns=200, cheese_count=21),
         )
 
@@ -125,7 +96,7 @@ class TestCreateBatch:
             batch_dir = create_batch(
                 parent_dir=tmpdir,
                 checkpoint_path=None,
-                mcts_config=PriorSamplingConfig(simulations=100),
+                mcts_config=DecoupledPUCTConfig(simulations=100),
                 game_params=GameParams(width=10, height=10, max_turns=200, cheese_count=21),
             )
 
@@ -159,7 +130,7 @@ class TestCreateBatch:
             batch_dir = create_batch(
                 parent_dir=tmpdir,
                 checkpoint_path=None,
-                mcts_config=PriorSamplingConfig(simulations=100),
+                mcts_config=DecoupledPUCTConfig(simulations=100),
                 game_params=GameParams(width=10, height=10, max_turns=200, cheese_count=21),
             )
 
@@ -180,7 +151,7 @@ class TestCreateBatch:
             batch_dir = create_batch(
                 parent_dir=tmpdir,
                 checkpoint_path=None,
-                mcts_config=PriorSamplingConfig(simulations=100),
+                mcts_config=DecoupledPUCTConfig(simulations=100),
                 game_params=GameParams(width=10, height=10, max_turns=200, cheese_count=21),
             )
 
@@ -191,31 +162,6 @@ class TestCreateBatch:
 
 class TestSaveLoadRoundtrip:
     """Tests for save/load metadata roundtrip."""
-
-    def test_roundtrip_prior_sampling(self) -> None:
-        """Metadata with PriorSamplingConfig survives roundtrip."""
-        with tempfile.TemporaryDirectory() as tmpdir:
-            batch_dir = Path(tmpdir) / "test-batch"
-            batch_dir.mkdir()
-
-            original = BatchMetadata(
-                batch_id="test-batch",
-                created_at=datetime(2024, 6, 15, 12, 30, 0, tzinfo=UTC),
-                checkpoint_path=None,
-                mcts_config=PriorSamplingConfig(simulations=800, gamma=0.95),
-                game_params=GameParams(width=20, height=15, max_turns=400, cheese_count=21),
-            )
-
-            save_batch_metadata(batch_dir, original)
-            loaded = load_batch_metadata(batch_dir)
-
-            assert loaded.batch_id == original.batch_id
-            assert loaded.created_at == original.created_at
-            assert loaded.checkpoint_path == original.checkpoint_path
-            assert isinstance(loaded.mcts_config, PriorSamplingConfig)
-            assert loaded.mcts_config.simulations == 800
-            assert loaded.mcts_config.gamma == 0.95
-            assert loaded.game_params.width == 20
 
     def test_roundtrip_decoupled_puct(self) -> None:
         """Metadata with DecoupledPUCTConfig survives roundtrip."""
@@ -236,7 +182,10 @@ class TestSaveLoadRoundtrip:
 
             assert loaded.checkpoint_path == "/models/best.pt"
             assert isinstance(loaded.mcts_config, DecoupledPUCTConfig)
+            assert loaded.mcts_config.simulations == 400
+            assert loaded.mcts_config.gamma == 0.98
             assert loaded.mcts_config.c_puct == 2.5
+            assert loaded.game_params.width == 10
 
 
 class TestGetBatchStats:
